feat: sepearte kv store and wal
